MCR-2110 Cleanup code
run IntelliJ "Silent Code Cleanup"

+++ mycore-base/src/main/java/org/mycore/frontend/servlets/MCRDerivateContentTransformerServlet.java
@@ -47,7 +47,7 @@ public class MCRDerivateContentTransformerServlet extends MCRContentServlet {
 
         FileTime lastModifiedTime = Files.getLastModifiedTime(mcrPath);
 
-        MCRFrontendUtil.writeCacheHeaders(resp, (long) CACHE_TIME, lastModifiedTime.toMillis(), true);
+        MCRFrontendUtil.writeCacheHeaders(resp, CACHE_TIME, lastModifiedTime.toMillis(), true);
 
         try {
             return getLayoutService().getTransformedContent(req, resp, pc);
